Added swipe spinning
Added swipe spinning. Swiping with the mosue will cause the cube to spin
at a velocity depending on the velocity of the mouse. No friction added
yet, clicking again with the mouse will abruptly stop the cube.

-Added startSpeedometer function that starts a looping function that
polls the mouse position ever few milliseconds to determine the average
velocity of the mouse. On mouseup, this velocity is added to the cube.
Sensitivity of the speedometer can be changed by speedSensitivity

-made global variables for constants like
  framesPerSecond: for converting degrees per second into degrees per 
  frame 

  speedSensitivity: interval of velocity polling

  rotationScale: multiplied by actual pixel movement of mouse to get
  desired number of degrees of cube rotation

-Changed face colors to match rubix cube

-rotateCube returns the new TransformMatrix after it's done. That way,
the caller can decide if they want to do anything with it or not.

-reintroduced mousePos STRef. Used with speedometer.

-using pattern matching instead of if else for some functions in
Matrices module

-added functions to Matrices module
  sum: takes an array of RotationVectors and adds them using vector
  addition.

  average: takes an array of RotationVectors and finds the average by
  calling sum, dividing the magnitude by the number of vectors, and then
  recalculating the x and y components from the new magnitude.